% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{NNKCDE package}
\alias{NNKCDE package}
\alias{NNKCDE package-package}
\title{Nearest Neighboor Kernel Conditional Density Estimation}
\description{
A package to compute nearest neighbor kernel conditional density
estimates. Provides functions for prediction and tuning based upon
conditional density loss
}
